+++ core/model/auth.go
@@ -21,6 +21,7 @@ import (
 
 	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth"
 	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth/authorization"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/errors"
 	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logutils"
 )
 
@@ -206,12 +207,56 @@ func (ls LoginSession) isYearlyExpired(policy YearlyExpirePolicy) bool {
 }
 
 // CurrentRefreshToken returns the current refresh token (last element of RefreshTokens)
-func (ls LoginSession) CurrentRefreshToken() string {
+func (ls LoginSession) CurrentRefreshToken() (string, error) {
 	numTokens := len(ls.RefreshTokens)
 	if numTokens <= 0 {
-		return ""
+		return "", errors.ErrorData(logutils.StatusMissing, TypeRefreshToken, &logutils.FieldArgs{"current": true})
 	}
-	return ls.RefreshTokens[numTokens-1]
+
+	currentToken := ls.RefreshTokens[numTokens-1]
+	if currentToken == "" {
+		return "", errors.ErrorData("Empty", TypeRefreshToken, &logutils.FieldArgs{"current": true})
+	}
+
+	return currentToken, nil
+}
+
+// PreviousRefreshToken returns the previous refresh token (second to last element of RefreshTokens)
+func (ls LoginSession) PreviousRefreshToken() (string, error) {
+	numTokens := len(ls.RefreshTokens)
+	if numTokens <= 1 {
+		return "", errors.ErrorData(logutils.StatusMissing, TypeRefreshToken, &logutils.FieldArgs{"current": false})
+	}
+
+	previousToken := ls.RefreshTokens[numTokens-2]
+	if previousToken == "" {
+		return "", errors.ErrorData("Empty", TypeRefreshToken, &logutils.FieldArgs{"current": false})
+	}
+
+	return previousToken, nil
+}
+
+// IsInRefreshGracePeriod return whether the login session is in refresh grace period
+func (ls LoginSession) IsInRefreshGracePeriod(now *time.Time) bool {
+	loginsSessionsSetting := ls.AppOrg.LoginsSessionsSetting
+
+	gracePeriodPolicy := loginsSessionsSetting.RefreshGracePeriodPolicy
+	if !gracePeriodPolicy.Active {
+		return false
+	}
+
+	lastRefreshTime := ls.DateCreated
+	if ls.DateRefreshed != nil {
+		lastRefreshTime = *ls.DateRefreshed //previsouly refreshed
+	}
+
+	if now == nil {
+		currentTime := time.Now()
+		now = &currentTime
+	}
+	gracePeriodDuration := time.Duration(gracePeriodPolicy.GracePeriod) * time.Second
+	gracePeriodEndTime := lastRefreshTime.Add(gracePeriodDuration)
+	return now.Before(gracePeriodEndTime)
 }
 
 // LogInfo gives the information appropriate to be logged for the session